seprate project list, add
