+++ src/components/Project/AddProject.jsx
@@ -1,15 +1,5 @@
 import { useState, useEffect } from "react"
-import Client from "../services/api"
-
-export const GetProjects = async () => {
-  try {
-    const res = await Client.get("/projects/project")
-    return res.data
-  } catch (error) {
-    console.error("Error fetching projects:", error)
-    throw error
-  }
-}
+import Client from "../../services/api"
 
 export const createProject = async (projectData) => {
   try {
@@ -24,14 +14,12 @@ export const createProject = async (projectData) => {
   }
 }
 
-const Project = () => {
+const AddProject = () => {
   const [title, setTitle] = useState("")
   const [description, setDescription] = useState("")
   const [service, setService] = useState("")
   const [userId, setUserId] = useState("")
   const [files, setFiles] = useState([])
-  const [projects, setProjects] = useState([])
-  const [message, setMessage] = useState("")
 
   useEffect(() => {
     loadProjects()
@@ -67,43 +55,6 @@ const Project = () => {
     } catch (error) {
       setMessage("Error creating project")
       console.error(error)
-    }
-  }
-
-  const renderFile = (fileUrl) => {
-    // Prefix the file path with the backend URL
-    const adjustedUrl = `http://localhost:4000/${fileUrl.replace(
-      /^public[\\/]+/,
-      ""
-    )}`
-
-    const fileExtension = adjustedUrl.split(".").pop().toLowerCase()
-
-    if (["jpg", "jpeg", "png", "gif"].includes(fileExtension)) {
-      return (
-        <img
-          src={adjustedUrl}
-          alt="project-file"
-          style={{ maxWidth: "40%", height: "40%" }}
-        />
-      )
-    } else if (fileExtension === "pdf") {
-      return (
-        <iframe
-          src={adjustedUrl}
-          type="application/pdf"
-          width="100%"
-          height="500px"
-          style={{ border: "none" }}
-          title="PDF Preview"
-        ></iframe>
-      )
-    } else {
-      return (
-        <a href={adjustedUrl} target="_blank" rel="noopener noreferrer">
-          {fileUrl}
-        </a>
-      )
     }
   }
 
@@ -152,35 +103,8 @@ const Project = () => {
         </div>
         <button type="submit">Create Project</button>
       </form>
-
-      {message && <p>{message}</p>}
-
-      <h3>All Projects</h3>
-      {projects.map((project) => (
-        <div
-          key={project._id}
-          style={{
-            border: "1px solid #ccc",
-            padding: "10px",
-            margin: "10px 0",
-          }}
-        >
-          <h3>{project.title}</h3>
-          <p>{project.description}</p>
-          <p>Service: {project.service?.name}</p>
-          <p>User: {project.user?.username}</p>
-          <div>
-            <strong>Files:</strong>
-            {project.files.map((file, index) => (
-              <div key={index} style={{ marginBottom: "10px" }}>
-                {renderFile(file)}
-              </div>
-            ))}
-          </div>
-        </div>
-      ))}
     </div>
   )
 }
 
-export default Project
+export default AddProject
